refactoring controllers items
