+++ models/contacts.json
@@ -1,17 +1,5 @@
 [
   {
-    "id": "3",
-    "name": "423423ull5558586585545a",
-    "email": "423423432423423@Nulla.com",
-    "phone": "12345678"
-  },
-  {
-    "id": "6",
-    "name": "Abbot Franks",
-    "email": "scelerisque@magnis.org",
-    "phone": "(186) 568-3720"
-  },
-  {
     "id": "7",
     "name": "Reuben Henry",
     "email": "pharetra.ut@dictum.co.uk",
@@ -36,27 +24,21 @@
     "phone": "(748) 206-2688"
   },
   {
-    "id": "0c30b961b12",
-    "name": "dasdsdd",
-    "email": "423423432423423@Nulla.com",
-    "phone": "7043432423"
+    "id": "77787c45b2c",
+    "name": "8888888888",
+    "email": "66666@Nulla.com",
+    "phone": "1233166114"
   },
   {
-    "id": "b1b023f28ee",
-    "name": "dds1212dd",
-    "email": "423423432423423@Nulla.com",
-    "phone": "4555045523"
+    "id": "7787c45b2c9",
+    "name": "8888888888",
+    "email": "66666@Nulla.com",
+    "phone": "1233166114"
   },
   {
-    "id": "1b023f28eeb",
-    "name": "dds1212dd",
-    "email": "423423432423423@Nulla.com",
-    "phone": "12345678"
-  },
-  {
-    "id": "e1c91eb5199",
-    "name": "Howard",
-    "email": "423423432423423@Nulla.com",
-    "phone": "725272704"
+    "id": "bdb4f55d506",
+    "name": "000000000000",
+    "email": "00000000000@Nulla.com",
+    "phone": "00000000"
   }
 ]
