+++ models/index.js
@@ -1,0 +1,7 @@
+const getContacts = require("./getContacts");
+const writeContacts = require("./writeContacts");
+
+module.exports = {
+  getContacts,
+  writeContacts,
+};
