+++ routes/api/contacts.js
@@ -1,14 +1,15 @@
 const express = require("express");
+
 const {
   listContacts,
   getContactById,
   removeContact,
   addContact,
   updateContact,
-} = require("../../controllers/contactsController");
+} = require("../../controllers");
 
 const {
-  addAndUpdateContactValidation,
+  addAndUpdateValidation,
   isValidId,
 } = require("../../middlewares/validationMiddleware");
 
@@ -18,15 +19,10 @@ router.get("/", listContacts);
 
 router.get("/:contactId", isValidId, getContactById);
 
-router.post("/", addAndUpdateContactValidation, addContact);
+router.post("/", addAndUpdateValidation, addContact);
 
 router.delete("/:contactId", isValidId, removeContact);
 
-router.put(
-  "/:contactId",
-  addAndUpdateContactValidation,
-  isValidId,
-  updateContact
-);
+router.put("/:contactId", addAndUpdateValidation, isValidId, updateContact);
 
 module.exports = { contactsRouter: router };
