+++ server.js
@@ -1,5 +1,7 @@
 const app = require("./app");
 
-app.listen(3000, () => {
+const { PORT = 3000 } = process.env;
+
+app.listen(PORT, () => {
   console.log("Server running. Use our API on port: 3000");
 });
